Updated seed generation

+++ examples/hpc/barren_plateau_trial.jl
@@ -18,7 +18,7 @@ function run_trial(config::Dict{Symbol, Any}, trial_id)
     h = config[:h]
     g = config[:g]
 
-    seed = Random.rand(Int)
+    seed = Int(Random.rand(UInt16))
     results[:seed] = seed
     Random.seed!(seed)
